Changed yajl to parse a simplified json format, easier to edit by hand. added GetChildByName in the SceneNode, will stop at the first match if there is doubles. Need improve this, but works for now. Changed Viewport and ViewManager to use u32. Updated sdl mouse input to return raw screen coords. Minor changes on image render. Added Camera to scene object factory, so we can load N cameras in the scene. Added an image sample that uses mouse to move an image (it is not moving correctly.. fix it). Added a sample of rendering configuration via json, do some tests and implement an initial version to test it.

+++ include/ViewManager.h
@@ -54,7 +54,7 @@ class SEED_CORE_API ViewManager : public IModule
 		virtual Renderer *GetCurrentRenderer() const;
 		virtual Viewport *GetCurrentViewport() const;
 
-		virtual Viewport *GetViewportAt(f32 x, f32 y);
+		virtual Viewport *GetViewportAt(u32 x, u32 y);
 
 		// IModule
 		virtual bool Initialize();
